SYLIUS-248 refactor SupportedRefundPaymentMethodsProviderDecorators
