+++ src/Provider/AmexSupportedRefundPaymentMethodsProviderDecorator.php
@@ -5,107 +5,9 @@ declare(strict_types=1);
 namespace PayPlug\SyliusPayPlugPlugin\Provider;
 
 use PayPlug\SyliusPayPlugPlugin\Gateway\AmericanExpressGatewayFactory;
-use Payum\Core\Model\GatewayConfigInterface;
-use Sylius\Component\Core\Model\ChannelInterface;
-use Sylius\Component\Core\Model\OrderInterface;
-use Sylius\Component\Core\Model\PaymentInterface;
-use Sylius\Component\Core\Model\PaymentMethodInterface;
-use Sylius\Component\Core\Repository\OrderRepositoryInterface;
 use Sylius\RefundPlugin\Provider\RefundPaymentMethodsProviderInterface;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\RequestStack;
 
-final class AmexSupportedRefundPaymentMethodsProviderDecorator implements RefundPaymentMethodsProviderInterface
+final class AmexSupportedRefundPaymentMethodsProviderDecorator extends AbstractSupportedRefundPaymentMethodsProvider implements RefundPaymentMethodsProviderInterface
 {
-    /** @var RefundPaymentMethodsProviderInterface */
-    private $decorated;
-
-    /** @var RequestStack */
-    private $requestStack;
-
-    /** @var OrderRepositoryInterface */
-    private $orderRepository;
-
-    public function __construct(
-        RefundPaymentMethodsProviderInterface $decorated,
-        RequestStack $requestStack,
-        OrderRepositoryInterface $orderRepository
-    ) {
-        $this->decorated = $decorated;
-        $this->requestStack = $requestStack;
-        $this->orderRepository = $orderRepository;
-    }
-
-    public function findForChannel(ChannelInterface $channel): array
-    {
-        $paymentMethods = $this->decorated->findForChannel($channel);
-        $request = $this->requestStack->getCurrentRequest();
-        if (!$request instanceof Request || 'sylius_refund_order_refunds_list' !== $request->get('_route')) {
-            return $paymentMethods;
-        }
-
-        $orderNumber = $request->get('orderNumber');
-        if (!is_string($orderNumber)) {
-            return $paymentMethods;
-        }
-
-        /** @var OrderInterface|null $order */
-        $order = $this->orderRepository->findOneByNumber($orderNumber);
-
-        if (!$order instanceof OrderInterface) {
-            return $paymentMethods;
-        }
-
-        if ($this->isAmericanExpress($order)) {
-            return array_filter($paymentMethods, function (PaymentMethodInterface $paymentMethod) use ($order): bool {
-                $lastPayment = $order->getLastPayment(PaymentInterface::STATE_COMPLETED);
-                if (!$lastPayment instanceof PaymentInterface) {
-                    return false;
-                }
-                $lastPaymentMethod = $lastPayment->getMethod();
-                if (!$lastPaymentMethod instanceof PaymentMethodInterface) {
-                    return false;
-                }
-
-                return $paymentMethod->getId() === $lastPaymentMethod->getId();
-            });
-        }
-
-        foreach ($paymentMethods as $key => $paymentMethod) {
-            /** @var GatewayConfigInterface $gatewayConfig */
-            $gatewayConfig = $paymentMethod->getGatewayConfig();
-
-            if (AmericanExpressGatewayFactory::FACTORY_NAME !== $gatewayConfig->getFactoryName()) {
-                continue;
-            }
-            unset($paymentMethods[$key]);
-        }
-
-        return $paymentMethods;
-    }
-
-    private function isAmericanExpress(OrderInterface $order): bool
-    {
-        $lastPayment = $order->getLastPayment(PaymentInterface::STATE_COMPLETED);
-        if (!$lastPayment instanceof PaymentInterface) {
-            return false;
-        }
-
-        $paymentMethod = $lastPayment->getMethod();
-        if (!$paymentMethod instanceof PaymentMethodInterface) {
-            return false;
-        }
-
-        $gatewayConfig = $paymentMethod->getGatewayConfig();
-
-        if (!$gatewayConfig instanceof GatewayConfigInterface) {
-            return false;
-        }
-
-        if (AmericanExpressGatewayFactory::FACTORY_NAME !== $gatewayConfig->getFactoryName()) {
-            return false;
-        }
-
-        return true;
-    }
+    protected string $gatewayFactoryName = AmericanExpressGatewayFactory::FACTORY_NAME;
 }
