+++ src/Provider/OneySupportedRefundPaymentMethodsProviderDecorator.php
@@ -15,92 +15,7 @@ use Sylius\RefundPlugin\Provider\RefundPaymentMethodsProviderInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 
-final class OneySupportedRefundPaymentMethodsProviderDecorator implements RefundPaymentMethodsProviderInterface
+final class OneySupportedRefundPaymentMethodsProviderDecorator extends AbstractSupportedRefundPaymentMethodsProvider implements RefundPaymentMethodsProviderInterface
 {
-    /** @var RefundPaymentMethodsProviderInterface */
-    private $decorated;
-
-    /** @var RequestStack */
-    private $requestStack;
-
-    /** @var OrderRepositoryInterface */
-    private $orderRepository;
-
-    public function __construct(
-        RefundPaymentMethodsProviderInterface $decorated,
-        RequestStack $requestStack,
-        OrderRepositoryInterface $orderRepository
-    ) {
-        $this->decorated = $decorated;
-        $this->requestStack = $requestStack;
-        $this->orderRepository = $orderRepository;
-    }
-
-    public function findForChannel(ChannelInterface $channel): array
-    {
-        $paymentMethods = $this->decorated->findForChannel($channel);
-        $request = $this->requestStack->getCurrentRequest();
-        if (!$request instanceof Request || 'sylius_refund_order_refunds_list' !== $request->get('_route')) {
-            return $paymentMethods;
-        }
-
-        /** @var OrderInterface|null $order */
-        $order = $this->orderRepository->findOneByNumber($request->get('orderNumber'));
-
-        if (!$order instanceof OrderInterface) {
-            return $paymentMethods;
-        }
-
-        if ($this->isOneyPayment($order)) {
-            return array_filter($paymentMethods, function (PaymentMethodInterface $paymentMethod) use ($order): bool {
-                $lastPayment = $order->getLastPayment(PaymentInterface::STATE_COMPLETED);
-                if (!$lastPayment instanceof PaymentInterface) {
-                    return false;
-                }
-                $lastPaymentMethod = $lastPayment->getMethod();
-                if (!$lastPaymentMethod instanceof PaymentMethodInterface) {
-                    return false;
-                }
-
-                return $paymentMethod->getId() === $lastPaymentMethod->getId();
-            });
-        }
-
-        foreach ($paymentMethods as $key => $paymentMethod) {
-            /** @var GatewayConfigInterface $gatewayConfig */
-            $gatewayConfig = $paymentMethod->getGatewayConfig();
-
-            if (OneyGatewayFactory::FACTORY_NAME !== $gatewayConfig->getFactoryName()) {
-                continue;
-            }
-            unset($paymentMethods[$key]);
-        }
-
-        return $paymentMethods;
-    }
-
-    private function isOneyPayment(OrderInterface $order): bool
-    {
-        $lastPayment = $order->getLastPayment(PaymentInterface::STATE_COMPLETED);
-        if (!$lastPayment instanceof PaymentInterface) {
-            return false;
-        }
-
-        $paymentMethod = $lastPayment->getMethod();
-        if (!$paymentMethod instanceof PaymentMethodInterface) {
-            return false;
-        }
-
-        $gatewayConfig = $paymentMethod->getGatewayConfig();
-
-        if (!$gatewayConfig instanceof GatewayConfigInterface) {
-            return false;
-        }
-
-        if (OneyGatewayFactory::FACTORY_NAME !== $gatewayConfig->getFactoryName()) {
-            return false;
-        }
-
-        return true;
-    }
+    protected string $gatewayFactoryName = OneyGatewayFactory::FACTORY_NAME;
 }
