+++ src/Provider/SupportedRefundPaymentMethodsProviderDecorator.php
@@ -15,19 +15,11 @@ use Sylius\RefundPlugin\Provider\RefundPaymentMethodsProviderInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 
-final class SupportedRefundPaymentMethodsProviderDecorator implements RefundPaymentMethodsProviderInterface
+final class SupportedRefundPaymentMethodsProviderDecorator extends AbstractSupportedRefundPaymentMethodsProvider implements RefundPaymentMethodsProviderInterface
 {
-    /** @var RefundPaymentMethodsProviderInterface */
-    private $decorated;
+    protected array $supportedRefundGateways;
 
-    /** @var RequestStack */
-    private $requestStack;
-
-    /** @var OrderRepositoryInterface */
-    private $orderRepository;
-
-    /** @var array */
-    private $supportedRefundGateways;
+    protected string $gatewayFactoryName = PayPlugGatewayFactory::FACTORY_NAME;
 
     public function __construct(
         RefundPaymentMethodsProviderInterface $decorated,
@@ -35,28 +27,13 @@ final class SupportedRefundPaymentMethodsProviderDecorator implements RefundPaym
         OrderRepositoryInterface $orderRepository,
         array $supportedRefundGateways
     ) {
-        $this->decorated = $decorated;
-        $this->requestStack = $requestStack;
-        $this->orderRepository = $orderRepository;
+        parent::__construct($decorated, $requestStack, $orderRepository);
         $this->supportedRefundGateways = $supportedRefundGateways;
     }
 
-    public function findForChannel(ChannelInterface $channel): array
+    protected function find(array $paymentMethods, OrderInterface $order): array
     {
-        $paymentMethods = $this->decorated->findForChannel($channel);
-        $request = $this->requestStack->getCurrentRequest();
-        if (!$request instanceof Request || 'sylius_refund_order_refunds_list' !== $request->get('_route')) {
-            return $paymentMethods;
-        }
-
-        /** @var OrderInterface|null $order */
-        $order = $this->orderRepository->findOneByNumber($request->get('orderNumber'));
-
-        if (!$order instanceof OrderInterface) {
-            return $paymentMethods;
-        }
-
-        if ($this->isPayPlugPayment($order)) {
+        if ($this->isPayplugPayment($order)) {
             return array_filter($paymentMethods, function (PaymentMethodInterface $paymentMethod) use ($order): bool {
                 $lastPayment = $order->getLastPayment(PaymentInterface::STATE_COMPLETED);
                 if (!$lastPayment instanceof PaymentInterface) {
@@ -73,8 +50,8 @@ final class SupportedRefundPaymentMethodsProviderDecorator implements RefundPaym
 
         if (null !== $order->getLastPayment() &&
             null !== $order->getLastPayment()->getMethod() &&
-            PayPlugGatewayFactory::FACTORY_NAME === $order->getLastPayment()->getMethod()->getCode() &&
-            !\in_array(PayPlugGatewayFactory::FACTORY_NAME, $this->supportedRefundGateways, true)) {
+            $this->gatewayFactoryName === $order->getLastPayment()->getMethod()->getCode() &&
+            !\in_array($this->gatewayFactoryName, $this->supportedRefundGateways, true)) {
             $this->requestStack->getSession()->getFlashBag()->add('info', 'payplug_sylius_payplug_plugin.ui.payplug_refund_gateway_is_not_activated');
         }
 
@@ -82,37 +59,12 @@ final class SupportedRefundPaymentMethodsProviderDecorator implements RefundPaym
             /** @var GatewayConfigInterface $gatewayConfig */
             $gatewayConfig = $paymentMethod->getGatewayConfig();
 
-            if (PayPlugGatewayFactory::FACTORY_NAME !== $gatewayConfig->getFactoryName()) {
+            if ($this->gatewayFactoryName !== $gatewayConfig->getFactoryName()) {
                 continue;
             }
             unset($paymentMethods[$key]);
         }
 
         return $paymentMethods;
-    }
-
-    private function isPayPlugPayment(OrderInterface $order): bool
-    {
-        $lastPayment = $order->getLastPayment(PaymentInterface::STATE_COMPLETED);
-        if (!$lastPayment instanceof PaymentInterface) {
-            return false;
-        }
-
-        $paymentMethod = $lastPayment->getMethod();
-        if (!$paymentMethod instanceof PaymentMethodInterface) {
-            return false;
-        }
-
-        $gatewayConfig = $paymentMethod->getGatewayConfig();
-
-        if (!$gatewayConfig instanceof GatewayConfigInterface) {
-            return false;
-        }
-
-        if (PayPlugGatewayFactory::FACTORY_NAME !== $gatewayConfig->getFactoryName()) {
-            return false;
-        }
-
-        return true;
     }
 }
